+++ packages/runtime/src/index.ts
@@ -1,6 +1,6 @@
 import { join, relative } from 'path'
 
-import type { NetlifyPlugin } from '@netlify/build'
+import type { NetlifyPlugin, NetlifyPluginOptions } from '@netlify/build'
 import { bold, redBright } from 'chalk'
 import destr from 'destr'
 import { existsSync, readFileSync } from 'fs-extra'
@@ -17,13 +17,17 @@ import {
 } from './helpers/config'
 import { onPreDev } from './helpers/dev'
 import { writeEdgeFunctions, loadMiddlewareManifest, cleanupEdgeFunctions } from './helpers/edge'
-import { moveStaticPages, movePublicFiles, patchNextFiles } from './helpers/files'
+import { moveStaticPages, movePublicFiles, patchNextFiles, removeMetadataFiles } from './helpers/files'
+import { splitApiRoutes } from './helpers/flags'
 import {
   generateFunctions,
   setupImageFunction,
   generatePagesResolver,
-  getExtendedApiRouteConfigs,
   warnOnApiRoutes,
+  getAPILambdas,
+  packSingleFunction,
+  getExtendedApiRouteConfigs,
+  APILambda,
 } from './helpers/functions'
 import { generateRedirects, generateStaticRedirects } from './helpers/redirects'
 import { shouldSkip, isNextAuthInstalled, getCustomImageResponseHeaders, getRemotePatterns } from './helpers/utils'
@@ -72,7 +76,8 @@ const plugin: NetlifyPlugin = {
     utils: {
       build: { failBuild },
     },
-  }) {
+    featureFlags = {},
+  }: NetlifyPluginOptions & { featureFlags?: Record<string, unknown> }) {
     if (shouldSkip()) {
       return
     }
@@ -161,11 +166,22 @@ const plugin: NetlifyPlugin = {
 
     const buildId = readFileSync(join(publish, 'BUILD_ID'), 'utf8').trim()
 
-    await configureHandlerFunctions({ netlifyConfig, ignore, publish: relative(process.cwd(), publish) })
-    const apiRoutes = await getExtendedApiRouteConfigs(publish, appDir, pageExtensions)
+    const apiLambdas: APILambda[] = splitApiRoutes(featureFlags, publish)
+      ? await getAPILambdas(publish, appDir, pageExtensions)
+      : await getExtendedApiRouteConfigs(publish, appDir, pageExtensions).then((extendedRoutes) =>
+          extendedRoutes.map(packSingleFunction),
+        )
 
-    await generateFunctions(constants, appDir, apiRoutes)
+    await generateFunctions(constants, appDir, apiLambdas)
     await generatePagesResolver(constants)
+
+    await configureHandlerFunctions({
+      netlifyConfig,
+      ignore,
+      publish: relative(process.cwd(), publish),
+      apiLambdas,
+      splitApiRoutes: splitApiRoutes(featureFlags, publish),
+    })
 
     await movePublicFiles({ appDir, outdir, publish, basePath })
 
@@ -193,7 +209,7 @@ const plugin: NetlifyPlugin = {
       netlifyConfig,
       nextConfig: { basePath, i18n, trailingSlash, appDir },
       buildId,
-      apiRoutes,
+      apiLambdas,
     })
 
     await writeEdgeFunctions({ netlifyConfig, routesManifest })
@@ -238,6 +254,12 @@ const plugin: NetlifyPlugin = {
     warnForProblematicUserRewrites({ basePath, redirects })
     warnForRootRedirects({ appDir })
     await warnOnApiRoutes({ FUNCTIONS_DIST })
+
+    // we are removing metadata files from Publish directory
+    // we have to do this after functions were bundled as bundling still
+    // require those files, but we don't want to publish them
+    await removeMetadataFiles(publish)
+
     if (experimental?.appDir) {
       console.log(
         '🧪 Thank you for testing "appDir" support on Netlify. For known issues and to give feedback, visit https://ntl.fyi/next-13-feedback',